refactor keeper
