+++ modules/service/internal/keeper/binding.go
@@ -1,0 +1,251 @@
+package keeper
+
+import (
+	"fmt"
+	"time"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/irisnet/irishub/modules/service/internal/types"
+)
+
+func (k Keeper) AddServiceBinding(ctx sdk.Context, svcBinding types.SvcBinding) sdk.Error {
+	_, found := k.GetServiceDefinition(ctx, svcBinding.DefChainID, svcBinding.DefName)
+	if !found {
+		return types.ErrSvcDefNotExists(k.codespace, svcBinding.DefChainID, svcBinding.DefName)
+	}
+
+	_, found = k.GetServiceBinding(ctx, svcBinding.DefChainID, svcBinding.DefName, svcBinding.BindChainID, svcBinding.Provider)
+	if found {
+		return types.ErrSvcBindingExists(k.codespace)
+	}
+
+	minDeposit, err := k.getMinDeposit(ctx, svcBinding.Prices)
+	if err != nil {
+		return err
+	}
+
+	if !svcBinding.Deposit.IsAllGTE(minDeposit) {
+		return types.ErrLtMinProviderDeposit(k.Codespace(), minDeposit)
+	}
+
+	err = k.validateMethodPrices(ctx, svcBinding)
+	if err != nil {
+		return err
+	}
+
+	// Subtract coins from provider's account
+	err = k.bk.SendCoins(ctx, svcBinding.Provider, auth.ServiceDepositCoinsAccAddr, svcBinding.Deposit)
+	if err != nil {
+		return err
+	}
+
+	svcBinding.DisableTime = time.Time{}
+	k.SetServiceBinding(ctx, svcBinding)
+
+	return nil
+}
+
+func (k Keeper) SetServiceBinding(ctx sdk.Context, svcBinding types.SvcBinding) {
+	store := ctx.KVStore(k.storeKey)
+
+	bz := k.cdc.MustMarshalBinaryLengthPrefixed(svcBinding)
+	store.Set(types.GetServiceBindingKey(svcBinding.DefChainID, svcBinding.DefName, svcBinding.BindChainID, svcBinding.Provider), bz)
+}
+
+func (k Keeper) GetServiceBinding(ctx sdk.Context, defChainID, defName, bindChainID string, provider sdk.AccAddress) (svcBinding types.SvcBinding, found bool) {
+	store := ctx.KVStore(k.storeKey)
+
+	bz := store.Get(types.GetServiceBindingKey(defChainID, defName, bindChainID, provider))
+	if bz == nil {
+		return svcBinding, false
+	}
+
+	k.cdc.MustUnmarshalBinaryLengthPrefixed(bz, &svcBinding)
+	return svcBinding, true
+}
+
+func (k Keeper) ServiceBindingsIterator(ctx sdk.Context, defChainID, defName string) sdk.Iterator {
+	store := ctx.KVStore(k.storeKey)
+	return sdk.KVStorePrefixIterator(store, types.GetBindingsSubspaceKey(defChainID, defName))
+}
+
+func (k Keeper) UpdateServiceBinding(ctx sdk.Context, svcBinding types.SvcBinding) sdk.Error {
+	oldBinding, found := k.GetServiceBinding(ctx, svcBinding.DefChainID, svcBinding.DefName, svcBinding.BindChainID, svcBinding.Provider)
+	if !found {
+		return types.ErrSvcBindingNotExists(k.codespace)
+	}
+
+	if len(svcBinding.Prices) > 0 {
+		err := k.validateMethodPrices(ctx, svcBinding)
+		if err != nil {
+			return err
+		}
+
+		oldBinding.Prices = svcBinding.Prices
+	}
+
+	if svcBinding.BindingType != 0x00 {
+		oldBinding.BindingType = svcBinding.BindingType
+	}
+
+	// Add coins to svcBinding deposit
+	if !svcBinding.Deposit.IsAnyNegative() {
+		oldBinding.Deposit = oldBinding.Deposit.Add(svcBinding.Deposit)
+	}
+
+	// Subtract coins from provider's account
+	err := k.bk.SendCoins(ctx, svcBinding.Provider, auth.ServiceDepositCoinsAccAddr, svcBinding.Deposit)
+	if err != nil {
+		return err
+	}
+
+	if svcBinding.Level.UsableTime != 0 {
+		oldBinding.Level.UsableTime = svcBinding.Level.UsableTime
+	}
+	if svcBinding.Level.AvgRspTime != 0 {
+		oldBinding.Level.AvgRspTime = svcBinding.Level.AvgRspTime
+	}
+
+	// only check deposit if binding is available
+	if oldBinding.Available {
+		minDeposit, err := k.getMinDeposit(ctx, oldBinding.Prices)
+		if err != nil {
+			return err
+		}
+
+		if !oldBinding.Deposit.IsAllGTE(minDeposit) {
+			return types.ErrLtMinProviderDeposit(k.Codespace(), minDeposit.Sub(oldBinding.Deposit).Add(svcBinding.Deposit))
+		}
+	}
+
+	k.SetServiceBinding(ctx, oldBinding)
+
+	return nil
+}
+
+func (k Keeper) Disable(ctx sdk.Context, defChainID, defName, bindChainID string, provider sdk.AccAddress) sdk.Error {
+	binding, found := k.GetServiceBinding(ctx, defChainID, defName, bindChainID, provider)
+	if !found {
+		return types.ErrSvcBindingNotExists(k.codespace)
+	}
+
+	if !binding.Available {
+		return types.ErrDisable(k.Codespace(), "service binding is unavailable")
+	}
+
+	binding.Available = false
+	binding.DisableTime = ctx.BlockHeader().Time
+
+	k.SetServiceBinding(ctx, binding)
+
+	return nil
+}
+
+func (k Keeper) Enable(ctx sdk.Context, defChainID, defName, bindChainID string, provider sdk.AccAddress, deposit sdk.Coins) sdk.Error {
+	binding, found := k.GetServiceBinding(ctx, defChainID, defName, bindChainID, provider)
+	if !found {
+		return types.ErrSvcBindingNotExists(k.codespace)
+	}
+
+	if binding.Available {
+		return types.ErrEnable(k.Codespace(), "service binding is available")
+	}
+
+	// Add coins to svcBinding deposit
+	if !deposit.IsAnyNegative() {
+		binding.Deposit = binding.Deposit.Add(deposit)
+	}
+
+	minDeposit, err := k.getMinDeposit(ctx, binding.Prices)
+	if err != nil {
+		return err
+	}
+
+	if !binding.Deposit.IsAllGTE(minDeposit) {
+		return types.ErrLtMinProviderDeposit(k.codespace, minDeposit.Sub(binding.Deposit).Add(deposit))
+	}
+
+	// Subtract coins from provider's account
+	err = k.bk.SendCoins(ctx, binding.Provider, auth.ServiceDepositCoinsAccAddr, deposit)
+	if err != nil {
+		return err
+	}
+
+	binding.Available = true
+	binding.DisableTime = time.Time{}
+
+	k.SetServiceBinding(ctx, binding)
+
+	return nil
+}
+
+func (k Keeper) RefundDeposit(ctx sdk.Context, defChainID, defName, bindChainID string, provider sdk.AccAddress) sdk.Error {
+	binding, found := k.GetServiceBinding(ctx, defChainID, defName, bindChainID, provider)
+	if !found {
+		return types.ErrSvcBindingNotExists(k.Codespace())
+	}
+
+	if binding.Available {
+		return types.ErrRefundDeposit(k.Codespace(), "can't refund from a available service binding")
+	}
+
+	if binding.Deposit.IsZero() {
+		return types.ErrRefundDeposit(k.Codespace(), "service binding deposit is zero")
+	}
+
+	blockTime := ctx.BlockHeader().Time
+	params := k.GetParams(ctx)
+
+	refundTime := binding.DisableTime.Add(params.ArbitrationTimeLimit).Add(params.ComplaintRetrospect)
+	if blockTime.Before(refundTime) {
+		return types.ErrRefundDeposit(k.Codespace(), fmt.Sprintf("can not refund deposit before %s", refundTime.Format("2006-01-02 15:04:05")))
+	}
+
+	// Add coins to provider's account
+	err := k.bk.SendCoins(ctx, auth.ServiceDepositCoinsAccAddr, binding.Provider, binding.Deposit)
+	if err != nil {
+		return err
+	}
+
+	binding.Deposit = sdk.Coins{}
+	k.SetServiceBinding(ctx, binding)
+
+	return nil
+}
+
+func (k Keeper) getMinDeposit(ctx sdk.Context, prices []sdk.Coin) (sdk.Coins, sdk.Error) {
+	params := k.GetParams(ctx)
+	// min deposit must >= sum(method price) * minDepositMultiple
+	minDepositMultiple := sdk.NewInt(params.MinDepositMultiple)
+
+	var minDeposit sdk.Coins
+	for _, price := range prices {
+		if price.Amount.BigInt().BitLen()+minDepositMultiple.BigInt().BitLen()-1 > 255 {
+			return minDeposit, sdk.NewError(k.codespace, types.CodeIntOverflow, fmt.Sprintf("Int Overflow"))
+		}
+
+		minInt := price.Amount.Mul(minDepositMultiple)
+		minDeposit = minDeposit.Add(sdk.Coins{sdk.NewCoin(price.Denom, minInt)})
+	}
+
+	return minDeposit, nil
+}
+
+func (k Keeper) validateMethodPrices(ctx sdk.Context, svcBinding types.SvcBinding) sdk.Error {
+	iterator := k.GetMethods(ctx, svcBinding.DefChainID, svcBinding.DefName)
+	defer iterator.Close()
+
+	var methods []types.MethodProperty
+	for ; iterator.Valid(); iterator.Next() {
+		var method types.MethodProperty
+		k.cdc.MustUnmarshalBinaryLengthPrefixed(iterator.Value(), &method)
+		methods = append(methods, method)
+	}
+
+	if len(methods) != len(svcBinding.Prices) {
+		return types.ErrInvalidPriceCount(k.Codespace(), len(svcBinding.Prices), len(methods))
+	}
+
+	return nil
+}
