+++ modules/service/internal/types/expected_keepers.go
@@ -5,14 +5,12 @@ import (
 	"github.com/irisnet/irishub/modules/guardian"
 )
 
+// BankKeeper defines the expected bank keeper interface
 type BankKeeper interface {
-	AddCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, sdk.Error)
-
-	BurnCoins(ctx sdk.Context, fromAddr sdk.AccAddress, amt sdk.Coins) sdk.Error
-
 	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) sdk.Error
 }
 
+// GuardianKeeper defines the expected guardian keeper interface
 type GuardianKeeper interface {
 	GetTrustee(ctx sdk.Context, addr sdk.AccAddress) (guardian guardian.Guardian, found bool)
 }
